+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## WIP 0.8.0
 
+- Add `sort_keys` argument to `jsonreflow.dump()` and `jsonreflow.dumps()` ([#20](https://github.com/soxofaan/jsonreflow/issues/20))
+
 
 ## 0.7.1 - 2026-06-30